+++ critchecker/database.py
@@ -12,14 +12,14 @@ class Row():  # pylint: disable=too-many-instance-attributes
     A single row in a Critmas report.
 
     Args:
-        crit_parent_id: The parent deviation's ID.
-        crit_parent_type: The parent deviation's type ID.
-        crit_id: The critique's ID.
         crit_posted_at: The critique's timestamp.
         crit_edited_at: The critique's edited timestamp, if any.
         crit_author: The critique's author.
-        crit_words: The critique's length in words.
+        crit_words: Thw critique's length in words.
+        block_posted_at: The critique block's timestamp.
+        block_edited_at: The critique block's edited timestamp, if any.
         crit_url: The critique's URL.
+        block_url: The critique block's URL.
     """
 
     # pylint: disable=unsubscriptable-object
@@ -108,9 +108,29 @@ def format_timestamp(timestamp: str) -> str:
             "%Y-%m-%dT%H:%M:%S"
         )
     except (ValueError, TypeError) as exception:
-        raise ValueError(f"\"{timestamp}\": malformed timestamp") from exception
+        raise ValueError(f"'{timestamp}': malformed timestamp") from exception
 
     return as_datetime.strftime("%Y/%m/%d %H:%M")
+
+
+def timestamp_to_datetime(timestamp: str) -> datetime.datetime:
+    """
+    Parse a timestamp and return a datetime object.
+
+    Args:
+        timestamp: The timestamp to parse.
+
+    Returns:
+        The datetime obtained from parsing the timestamp.
+
+    Raises:
+        ValueError: If the timestamp is malformed.
+    """
+
+    try:
+        return datetime.datetime.strptime(timestamp, "%Y/%m/%d %H:%M")
+    except ValueError as exception:
+        raise ValueError(f"'{timestamp}': malformed timestamp") from exception
 
 
 def get_index_by_crit_url(url: str, data: list[Row]) -> int:
@@ -131,6 +151,6 @@ def get_index_by_crit_url(url: str, data: list[Row]) -> int:
     try:
         index = next(index for index, row in enumerate(data) if row.crit_url == url)
     except StopIteration as exception:
-        raise ValueError(f"\"{url}\" not found in database rows") from exception
+        raise ValueError(f"'{url}' not found in database rows") from exception
 
     return index