+++ critchecker/database.py
@@ -154,3 +154,30 @@ def get_index_by_crit_url(url: str, data: list[Row]) -> int:
         raise ValueError(f"'{url}' not found in database rows") from exception
 
     return index
+
+
+def measure_stats(data: list[Row]) -> tuple[int, int, int]:
+    """
+    Obtain the total, valid and deleted critiques from a database.
+
+    Args:
+        data: The database to parse.
+
+    Returns:
+        A tuple containing the amount of total, valid and deleted
+            critiques.
+
+    Raises:
+        ValueError: If the database contains invalid rows.
+    """
+
+    try:
+        # Let's use a generator to avoid copying the whole database.
+        deleted = sum((True for row in data if row.crit_words == 0))
+    except AttributeError as exception:
+        raise ValueError("invalid database") from exception
+
+    total = len(data)
+    valid = total - deleted
+
+    return total, valid, deleted